Remove minesweeper changes for separate PR
