+++ jumanji/environments/logic/minesweeper/constants.py
@@ -16,9 +16,8 @@ UNEXPLORED_ID: int = -1
 IS_MINE: int = 1
 PATCH_SIZE: int = 3
 REVEALED_EMPTY_SQUARE_REWARD: float = 1.0
-REVEALED_MINE_REWARD: float = 0.0
-INVALID_ACTION_REWARD: float = 0.0
-DEFAULT_COLOR_MAPPING: list = [
+REVEALED_MINE_OR_INVALID_ACTION_REWARD: float = 0.0
+COLOUR_MAPPING: list = [
     "orange",
     "blue",
     "green",
