+++ jumanji/environments/logic/minesweeper/env.py
@@ -12,34 +12,30 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import Optional, Sequence, Tuple
+from typing import List, Optional, Sequence, Tuple
 
 import chex
 import jax
 import jax.numpy as jnp
 import matplotlib.animation
+import matplotlib.pyplot as plt
 
+import jumanji.environments
 from jumanji import specs
 from jumanji.env import Environment
 from jumanji.environments.logic.minesweeper.constants import (
-    INVALID_ACTION_REWARD,
+    COLOUR_MAPPING,
     PATCH_SIZE,
-    REVEALED_EMPTY_SQUARE_REWARD,
-    REVEALED_MINE_REWARD,
     UNEXPLORED_ID,
 )
 from jumanji.environments.logic.minesweeper.done import DefaultDoneFn, DoneFn
-from jumanji.environments.logic.minesweeper.env_viewer import (
-    DefaultMinesweeperViewer,
-    MinesweeperViewer,
-)
-from jumanji.environments.logic.minesweeper.generator import (
-    Generator,
-    UniformSamplingGenerator,
-)
 from jumanji.environments.logic.minesweeper.reward import DefaultRewardFn, RewardFn
 from jumanji.environments.logic.minesweeper.types import Observation, State
-from jumanji.environments.logic.minesweeper.utils import count_adjacent_mines
+from jumanji.environments.logic.minesweeper.utils import (
+    count_adjacent_mines,
+    create_flat_mine_locations,
+    explored_mine,
+)
 from jumanji.types import TimeStep, restart, termination, transition
 
 
@@ -96,44 +92,46 @@ class Minesweeper(Environment[State]):
 
     def __init__(
         self,
+        num_rows: int = 10,
+        num_cols: int = 10,
+        num_mines: int = 10,
         reward_function: Optional[RewardFn] = None,
         done_function: Optional[DoneFn] = None,
-        env_viewer: Optional[MinesweeperViewer] = None,
-        generator: Optional[Generator] = None,
+        color_mapping: Optional[List[str]] = None,
     ):
         """Instantiate a `Minesweeper` environment.
 
         Args:
-
+            num_rows: number of rows, i.e. height of the board. Defaults to 10.
+            num_cols: number of columns, i.e. width of the board. Defaults to 10.
+            num_mines: number of mines on the board. Defaults to 10.
             reward_function: `RewardFn` whose `__call__` method computes the reward of an
                 environment transition based on the given current state and selected action.
                 Implemented options are [`DefaultRewardFn`]. Defaults to `DefaultRewardFn`.
             done_function: `DoneFn` whose `__call__` method computes the done signal given the
                 current state, action taken, and next state.
                 Implemented options are [`DefaultDoneFn`]. Defaults to `DefaultDoneFn`.
-            env_viewer: MinesweeperViewer to support rendering and animation methods.
-                Implemented options are [`DefaultMinesweeperViewer`].
-                Defaults to `DefaultMinesweeperViewer`.
-            generator: Generator to generate problem instances on environment reset.
-                Implemented options are [`SamplingGenerator`].
-                Defaults to `SamplingGenerator`.
-                The generator will have attributes:
-                    - num_rows: number of rows, i.e. height of the board. Defaults to 10.
-                    - num_cols: number of columns, i.e. width of the board. Defaults to 10.
-                    - num_mines: number of mines generated. Defaults to 10.
+            color_mapping: colour map used for rendering.
         """
-        self.reward_function = reward_function or DefaultRewardFn(
-            revealed_empty_square_reward=REVEALED_EMPTY_SQUARE_REWARD,
-            revealed_mine_reward=REVEALED_MINE_REWARD,
-            invalid_action_reward=INVALID_ACTION_REWARD,
-        )
+        if num_rows <= 1 or num_cols <= 1:
+            raise ValueError(
+                f"Should make a board of height and width greater than 1, "
+                f"got num_rows={num_rows}, num_cols={num_cols}"
+            )
+        if num_mines < 0 or num_mines >= num_rows * num_cols:
+            raise ValueError(
+                f"Number of mines should be constrained between 0 and the size of the board, "
+                f"got {num_mines}"
+            )
+        self.num_rows = num_rows
+        self.num_cols = num_cols
+        self.num_mines = num_mines
+        self.reward_function = reward_function or DefaultRewardFn()
         self.done_function = done_function or DefaultDoneFn()
-        self._generator = generator or UniformSamplingGenerator(
-            num_rows=10, num_cols=10, num_mines=10
-        )
-        self._env_viewer = env_viewer or DefaultMinesweeperViewer(
-            num_rows=self._generator.num_rows, num_cols=self._generator.num_cols
-        )
+
+        self.cmap = color_mapping if color_mapping else COLOUR_MAPPING
+        self.figure_name = f"{num_rows}x{num_cols} Minesweeper"
+        self.figure_size = (6.0, 6.0)
 
     def reset(self, key: chex.PRNGKey) -> Tuple[State, TimeStep[Observation]]:
         """Resets the environment.
@@ -146,7 +144,25 @@ class Minesweeper(Environment[State]):
             timestep: `TimeStep` corresponding to the first timestep returned by the
                 environment.
         """
-        state = self._generator(key)
+        key, sample_key = jax.random.split(key)
+        board = jnp.full(
+            shape=(self.num_rows, self.num_cols),
+            fill_value=UNEXPLORED_ID,
+            dtype=jnp.int32,
+        )
+        step_count = jnp.array(0, jnp.int32)
+        flat_mine_locations = create_flat_mine_locations(
+            key=sample_key,
+            num_rows=self.num_rows,
+            num_cols=self.num_cols,
+            num_mines=self.num_mines,
+        )
+        state = State(
+            board=board,
+            step_count=step_count,
+            key=key,
+            flat_mine_locations=flat_mine_locations,
+        )
         observation = self._state_to_observation(state=state)
         timestep = restart(observation=observation)
         return state, timestep
@@ -199,14 +215,14 @@ class Minesweeper(Environment[State]):
              - step_count: BoundedArray (int32) of shape ().
         """
         board = specs.BoundedArray(
-            shape=(self._generator.num_rows, self._generator.num_cols),
+            shape=(self.num_rows, self.num_cols),
             dtype=jnp.int32,
             minimum=-1,
             maximum=PATCH_SIZE * PATCH_SIZE - 1,
             name="board",
         )
         action_mask = specs.BoundedArray(
-            shape=(self._generator.num_rows, self._generator.num_cols),
+            shape=(self.num_rows, self.num_cols),
             dtype=bool,
             minimum=False,
             maximum=True,
@@ -216,14 +232,14 @@ class Minesweeper(Environment[State]):
             shape=(),
             dtype=jnp.int32,
             minimum=0,
-            maximum=self._generator.num_rows * self._generator.num_cols - 1,
+            maximum=self.num_rows * self.num_cols - 1,
             name="num_mines",
         )
         step_count = specs.BoundedArray(
             shape=(),
             dtype=jnp.int32,
             minimum=0,
-            maximum=self._generator.num_rows * self._generator.num_cols - self._generator.num_mines,
+            maximum=self.num_rows * self.num_cols - self.num_mines,
             name="step_count",
         )
         return specs.Spec(
@@ -243,9 +259,7 @@ class Minesweeper(Environment[State]):
             action_spec: `specs.MultiDiscreteArray` object.
         """
         return specs.MultiDiscreteArray(
-            num_values=jnp.array(
-                [self._generator.num_rows, self._generator.num_cols], jnp.int32
-            ),
+            num_values=jnp.array([self.num_rows, self.num_cols], jnp.int32),
             name="action",
             dtype=jnp.int32,
         )
@@ -254,20 +268,21 @@ class Minesweeper(Environment[State]):
         return Observation(
             board=state.board,
             action_mask=jnp.equal(state.board, UNEXPLORED_ID),
-            num_mines=jnp.array(
-                self._generator.num_mines, jnp.int32
-            ),
+            num_mines=jnp.array(self.num_mines, jnp.int32),
             step_count=state.step_count,
         )
 
-    def render(self, state: State, save_path: Optional[str] = None) -> None:
-        """Renders the current state of the board.
+    def render(self, state: State) -> None:
+        """Render the given environment state using matplotlib.
+
         Args:
-            state: the current state to be rendered.
-            save_path: the path where the image should be saved. If it is None, the plot
-                will not be saved.
+            state: environment state to be rendered.
+
         """
-        self._env_viewer.render(state=state)
+        self._clear_display()
+        fig, ax = self._get_fig_ax()
+        self._draw(ax, state)
+        self._update_display(fig)
 
     def animate(
         self,
@@ -275,22 +290,116 @@ class Minesweeper(Environment[State]):
         interval: int = 200,
         save_path: Optional[str] = None,
     ) -> matplotlib.animation.FuncAnimation:
-        """Creates an animated gif of the board based on the sequence of states.
-         Args:
-        states: a list of `State` objects representing the sequence of states.
-        interval: the delay between frames in milliseconds, default to 200.
-        save_path: the path where the animation file should be saved. If it is None, the plot
-            will not be saved.
+        """Create an animation from a sequence of environment states.
+
+        Args:
+            states: sequence of environment states corresponding to consecutive timesteps.
+            interval: delay between frames in milliseconds, default to 200.
+            save_path: the path where the animation file should be saved. If it is None, the plot
+                will not be saved.
+
         Returns:
-            animation.FuncAnimation: the animation object that was created.
+            Animation object that can be saved as a GIF, MP4, or rendered with HTML.
         """
-        return self._env_viewer.animate(
-            states=states, interval=interval, save_path=save_path
+        fig, ax = self._get_fig_ax()
+        plt.tight_layout()
+        plt.close(fig)
+
+        def make_frame(state_index: int) -> None:
+            state = states[state_index]
+            self._draw(ax, state)
+
+        # Create the animation object.
+        self._animation = matplotlib.animation.FuncAnimation(
+            fig,
+            make_frame,
+            frames=len(states),
+            interval=interval,
         )
+
+        # Save the animation as a GIF.
+        if save_path:
+            self._animation.save(save_path)
+
+        return self._animation
 
     def close(self) -> None:
         """Perform any necessary cleanup.
+
         Environments will automatically :meth:`close()` themselves when
         garbage collected or when the program exits.
         """
-        self._env_viewer.close()
+        plt.close(self.figure_name)
+
+    def _get_fig_ax(self) -> Tuple[plt.Figure, plt.Axes]:
+        exists = plt.fignum_exists(self.figure_name)
+        if exists:
+            fig = plt.figure(self.figure_name)
+            ax = fig.get_axes()[0]
+        else:
+            fig = plt.figure(self.figure_name, figsize=self.figure_size)
+            plt.suptitle(self.figure_name)
+            plt.tight_layout()
+            if not plt.isinteractive():
+                fig.show()
+            ax = fig.add_subplot()
+        return fig, ax
+
+    def _draw(self, ax: plt.Axes, state: State) -> None:
+        ax.clear()
+        ax.set_xticks(jnp.arange(-0.5, self.num_cols - 1, 1))
+        ax.set_yticks(jnp.arange(-0.5, self.num_rows - 1, 1))
+        ax.tick_params(
+            top=False,
+            bottom=False,
+            left=False,
+            right=False,
+            labelleft=False,
+            labelbottom=False,
+            labeltop=False,
+            labelright=False,
+        )
+        background = jnp.ones_like(state.board)
+        for i in range(self.num_rows):
+            for j in range(self.num_cols):
+                background = self._render_grid_square(
+                    state=state, ax=ax, i=i, j=j, background=background
+                )
+        ax.imshow(background, cmap="gray", vmin=0, vmax=1)
+        ax.grid(color="black", linestyle="-", linewidth=2)
+
+    def _render_grid_square(
+        self, state: State, ax: plt.Axes, i: int, j: int, background: chex.Array
+    ) -> chex.Array:
+        board_value = state.board[i, j]
+        if board_value != UNEXPLORED_ID:
+            if explored_mine(state=state, action=jnp.array([i, j], dtype=jnp.int32)):
+                background = background.at[i, j].set(0)
+            else:
+                ax.text(
+                    j,
+                    i,
+                    str(board_value),
+                    color=self.cmap[board_value],
+                    ha="center",
+                    va="center",
+                    fontsize="xx-large",
+                )
+        return background
+
+    def _update_display(self, fig: plt.Figure) -> None:
+        if plt.isinteractive():
+            # Required to update render when using Jupyter Notebook.
+            fig.canvas.draw()
+            if jumanji.environments.is_colab():
+                plt.show(self.figure_name)
+        else:
+            # Required to update render when not using Jupyter Notebook.
+            fig.canvas.draw_idle()
+            fig.canvas.flush_events()
+
+    def _clear_display(self) -> None:
+        if jumanji.environments.is_colab():
+            import IPython.display
+
+            IPython.display.clear_output(True)
