+++ jumanji/environments/logic/minesweeper/env_test.py
@@ -24,9 +24,8 @@ from jax import numpy as jnp
 from jax import random
 
 from jumanji.environments.logic.minesweeper.constants import (
-    INVALID_ACTION_REWARD,
     REVEALED_EMPTY_SQUARE_REWARD,
-    REVEALED_MINE_REWARD,
+    REVEALED_MINE_OR_INVALID_ACTION_REWARD,
 )
 from jumanji.environments.logic.minesweeper.env import Minesweeper
 from jumanji.environments.logic.minesweeper.types import State
@@ -70,12 +69,12 @@ def play_and_get_episode_stats(
         ),
         (
             [[0, 3], [0, 2]],
-            [REVEALED_EMPTY_SQUARE_REWARD, REVEALED_MINE_REWARD],
+            [REVEALED_EMPTY_SQUARE_REWARD, REVEALED_MINE_OR_INVALID_ACTION_REWARD],
             [StepType.MID, StepType.LAST],
         ),
         (
             [[0, 3], [0, 3]],
-            [REVEALED_EMPTY_SQUARE_REWARD, INVALID_ACTION_REWARD],
+            [REVEALED_EMPTY_SQUARE_REWARD, REVEALED_MINE_OR_INVALID_ACTION_REWARD],
             [StepType.MID, StepType.LAST],
         ),
     ],
@@ -108,11 +107,11 @@ def test_minesweeper_env_reset(minesweeper_env: Minesweeper) -> None:
     assert isinstance(timestep, TimeStep)
     assert isinstance(state, State)
     assert state.step_count == 0
-    assert state.flat_mine_locations.shape == (minesweeper_env._generator.num_mines,)
-    assert timestep.observation.num_mines == minesweeper_env._generator.num_mines
+    assert state.flat_mine_locations.shape == (minesweeper_env.num_mines,)
+    assert timestep.observation.num_mines == minesweeper_env.num_mines
     assert state.board.shape == (
-        minesweeper_env._generator.num_rows,
-        minesweeper_env._generator.num_cols,
+        minesweeper_env.num_rows,
+        minesweeper_env.num_cols,
     )
     assert jnp.array_equal(state.board, timestep.observation.board)
     assert timestep.observation.step_count == 0
@@ -190,9 +189,9 @@ def test_minesweeper_env_solved(minesweeper_env: Minesweeper) -> None:
     step_fn = jit(minesweeper_env.step)
     collected_rewards = []
     collected_step_types = []
-    for i in range(minesweeper_env._generator.num_rows):
-        for j in range(minesweeper_env._generator.num_cols):
-            flat_location = i * minesweeper_env._generator.num_cols + j
+    for i in range(minesweeper_env.num_rows):
+        for j in range(minesweeper_env.num_cols):
+            flat_location = i * minesweeper_env.num_cols + j
             if flat_location in state.flat_mine_locations:
                 continue
             action = jnp.array([i, j], dtype=jnp.int32)
@@ -200,8 +199,7 @@ def test_minesweeper_env_solved(minesweeper_env: Minesweeper) -> None:
             collected_rewards.append(timestep.reward)
             collected_step_types.append(timestep.step_type)
     expected_episode_length = (
-        minesweeper_env._generator.num_rows * minesweeper_env._generator.num_cols
-        - minesweeper_env._generator.num_mines
+        minesweeper_env.num_rows * minesweeper_env.num_cols - minesweeper_env.num_mines
     )
     assert collected_rewards == [REVEALED_EMPTY_SQUARE_REWARD] * expected_episode_length
     assert collected_step_types == [StepType.MID] * (expected_episode_length - 1) + [
